// web/generated/api.ts
export interface CreateProductDto {
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productTypeId': string;
    /**
     * 
     * @type {Array<CreateProductDtoAttributes>}
     * @memberof CreateProductDto
     */
    'attributes': Array<CreateProductDtoAttributes>;
}
/**
 * 
 * @export
 * @interface CreateProductDtoAttributes
 */
export interface CreateProductDtoAttributes {
    /**
     * 
     * @type {string}
     * @memberof CreateProductDtoAttributes
     */
    'name'?: string;
    /**
     * 
     * @type {Enum}
     * @memberof CreateProductDtoAttributes
     */
    'type'?: Enum;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDtoAttributes
     */
    'value'?: string;
}
/**
 * 
 * @export
 * @interface CreateProductTypeDto
 */
export interface CreateProductTypeDto {
    /**
     * 
     * @type {string}
     * @memberof CreateProductTypeDto
     */
    'name': string;
    /**
     * 
     * @type {Array<CreateProductTypeDtoAttributes>}
     * @memberof CreateProductTypeDto
     */
    'attributes': Array<CreateProductTypeDtoAttributes>;
}
/**
 * 
 * @export
 * @interface CreateProductTypeDtoAttributes
 */
export interface CreateProductTypeDtoAttributes {
    /**
     * 
     * @type {string}
     * @memberof CreateProductTypeDtoAttributes
     */
    'name'?: string;
    /**
     * 
     * @type {Enum}
     * @memberof CreateProductTypeDtoAttributes
     */
    'type'?: Enum;
}
/**
 * 
 * @export
 * @interface UpdateProductDto
 */
export interface UpdateProductDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateProductDto
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProductDto
     */
    'productTypeId'?: string;
    /**
     * 
     * @type {Array<CreateProductDtoAttributes>}
     * @memberof UpdateProductDto
     */
    'attributes'?: Array<CreateProductDtoAttributes>;
}
/**
 * 
 * @export
 * @interface UpdateProductTypeDto
 */
export interface UpdateProductTypeDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateProductTypeDto
     */
    'name'?: string;
    /**
     * 
     * @type {Array<CreateProductTypeDtoAttributes>}
     * @memberof UpdateProductTypeDto
     */
    'attributes'?: Array<CreateProductTypeDtoAttributes>;
}
// ...
export class ProductApi extends BaseAPI {
    /**
     * 
     * @param {CreateProductDto} createProductDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerCreate(createProductDto: CreateProductDto, options?: AxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerCreate(createProductDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerFindAll(options?: AxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerFindAll(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerFindOne(id: string, options?: AxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerFindOne(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerRemove(id: string, options?: AxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerRemove(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {UpdateProductDto} updateProductDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerUpdate(id: string, updateProductDto: UpdateProductDto, options?: AxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerUpdate(id, updateProductDto, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ProductTypeApi extends BaseAPI {
    /**
     * 
     * @param {CreateProductTypeDto} createProductTypeDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerCreate(createProductTypeDto: CreateProductTypeDto, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerCreate(createProductTypeDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerFindAll(options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerFindAll(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerFindOne(id: string, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerFindOne(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerRemove(id: string, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerRemove(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {UpdateProductTypeDto} updateProductTypeDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerUpdate(id: string, updateProductTypeDto: UpdateProductTypeDto, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerUpdate(id, updateProductTypeDto, options).then((request) => request(this.axios, this.basePath));
    }
}